Create GitHub workflow to publish nuget package (#3)
## Create GitHub workflow to publish nuget package

* Create build.yml file
* Restructure folders
* Moved packages to src folder
* Moved unit test to tests folder
* Change EF provider options to InMemory to fix failed unit test.

Issue #2
