+++ .github/workflows/build.yml
@@ -1,0 +1,48 @@
+name: Release
+
+on:
+  push:
+    paths:
+      - 'ATag.Core/**'
+      - 'ATag.EntityFrameworkCore/**'
+      - '**/build.yml'
+    branches:
+      - master
+  pull_request:
+    paths:
+      - 'ATag.Core/**'
+      - 'ATag.EntityFrameworkCore/**'
+      - '**/build.yml'
+  workflow_dispatch:
+
+jobs:
+  publish:
+    name: Build and Publish NuGets
+    timeout-minutes: 60
+    strategy:
+    # max-parallel: 2
+      matrix: 
+        os: ['windows-latest']
+    runs-on: ${{ matrix.os }}
+    if: github.ref == 'refs/heads/master' && success()
+    steps:
+      - name: Check out code
+        uses: actions/checkout@v2
+      - name: Install .NET Core SDK
+        uses: actions/setup-dotnet@v1
+        with:
+          dotnet-version: '3.1.x'
+      - name: Install .NET Core SDK
+        uses: actions/setup-dotnet@v1
+        with:
+          dotnet-version: '2.1.x'
+      - name: Clean
+        run: dotnet clean --configuration Release
+      - name: Build
+        run: dotnet build --configuration Release
+      - name: Test
+        run: dotnet test --configuration Release --no-restore --no-build -l trx --verbosity normal
+      - name: Pack
+        run: dotnet pack --no-build --configuration Release --output .
+      - name: PushNuget
+        run: dotnet nuget push *.nupkg --source https://api.nuget.org/v3/index.json --api-key ${{secrets.NUGET_API_KEY}} --skip-duplicate
