+++ .gitignore
@@ -59,3 +59,61 @@ node_modules
 /packages/*
 !/packages/repositories.config
 !NuGet.exe
+
+# MacOS
+**/.DS_Store
+
+### Rider ###
+# Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio, WebStorm and Rider
+# Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
+
+# User-specific stuff
+**/*.idea/**/workspace.xml
+**/*.idea/**/tasks.xml
+**/*.idea/**/usage.statistics.xml
+**/*.idea/**/dictionaries
+**/*.idea/**/shelf
+**/*.idea/**/misc.xml
+
+# AWS User-specific
+**/*.idea/**/aws.xml
+
+# SonarLint
+**/*.idea/sonarlint/
+
+# Generated files
+**/*.idea/**/contentModel.xml
+
+# Sensitive or high-churn files
+**/*.idea/**/dataSources/
+**/*.idea/**/dataSources.ids
+**/*.idea/**/dataSources.local.xml
+**/*.idea/**/sqlDataSources.xml
+**/*.idea/**/dynamic.xml
+**/*.idea/**/uiDesigner.xml
+**/*.idea/**/dbnavigator.xml
+
+# Gradle
+**/*.idea/**/gradle.xml
+**/*.idea/**/libraries
+
+# Mongo Explorer plugin
+**/*.idea/**/mongoSettings.xml
+
+# File-based project format
+*.iws
+
+# IntelliJ
+out/
+
+# mpeltonen/sbt-idea plugin
+**/*.idea_modules/
+
+# Cursive Clojure plugin
+**/*.idea/replstate.xml
+
+# Editor-based Rest Client
+**/*.idea/httpRequests
+
+# Android studio 3.1+ serialized cache file
+**/*.idea/caches/build_file_checksums.ser
