+++ ATag.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26430.15
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATag.Core", "ATag.Core\ATag.Core.csproj", "{5983565D-5B5E-4809-9DE3-E8E79470C401}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATag.Core", "src\ATag.Core\ATag.Core.csproj", "{5983565D-5B5E-4809-9DE3-E8E79470C401}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{3194B4A9-EC78-4149-BEA8-B9487EB43A16}"
 	ProjectSection(SolutionItems) = preProject
@@ -12,9 +12,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATag.EntityFrameworkCore", "ATag.EntityFrameworkCore\ATag.EntityFrameworkCore.csproj", "{B02A5C78-D33F-4D24-A166-121764B99006}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATag.EntityFrameworkCore", "src\ATag.EntityFrameworkCore\ATag.EntityFrameworkCore.csproj", "{B02A5C78-D33F-4D24-A166-121764B99006}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATag.Tests", "ATag.Tests\ATag.Tests.csproj", "{D4DC111F-5DB2-4544-94D3-C6D899CA1A52}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATag.Tests", "tests\ATag.Tests\ATag.Tests.csproj", "{D4DC111F-5DB2-4544-94D3-C6D899CA1A52}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{931DE703-8386-40F5-AE36-BB97DF5AC185}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -37,5 +39,8 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{D4DC111F-5DB2-4544-94D3-C6D899CA1A52} = {931DE703-8386-40F5-AE36-BB97DF5AC185}
 	EndGlobalSection
 EndGlobal
