+++ src/ATag.Core/ATag.Core.csproj
