+++ src/ATag.Core/Extensions.cs
