+++ src/ATag.Core/PagedEntity.cs
