+++ src/ATag.Core/Tag.cs
