+++ src/ATag.Core/TagException.cs
