+++ src/ATag.Core/TagNote.cs
