+++ src/ATag.Core/TaggedEntity.cs
