+++ src/ATag.EntityFrameworkCore/ATag.EntityFrameworkCore.csproj
@@ -1,0 +1,43 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFrameworks>netstandard1.4;netstandard2.0</TargetFrameworks>
+        <Authors>UNOPS</Authors>
+        <Version>0.0.14-beta</Version>
+        <Company>UNOPS</Company>
+        <Description>EF Core target for ATag</Description>
+        <PackageProjectUrl>https://github.com/UNOPS/atag</PackageProjectUrl>
+        <RepositoryUrl>https://github.com/UNOPS/atag</RepositoryUrl>
+        <AssemblyVersion>1.0.0.0</AssemblyVersion>
+        <PackageTags>tag label category</PackageTags>
+        <Copyright>Copyright 2017 (c) UNOPS. All rights reserved.</Copyright>
+        <DocumentationFile>bin\$(Configuration)\$(Platform)\ATag.EntityFrameworkCore.xml</DocumentationFile>
+        <RepositoryType>Github</RepositoryType>
+        <PackageLicenseUrl>https://github.com/UNOPS/atag/blob/develop/LICENSE</PackageLicenseUrl>
+        <NeutralLanguage>en-US</NeutralLanguage>
+        <PackageReleaseNotes>Added support for .NET Core 2.0</PackageReleaseNotes>
+        <PackageIconUrl>https://github.com/UNOPS/atag/blob/develop/atag.png</PackageIconUrl>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="LinqKit.Microsoft.EntityFrameworkCore" Version="1.1.10"/>
+    </ItemGroup>
+
+    <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.0.0"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.0.0"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="2.0.0"/>
+    </ItemGroup>
+
+    <ItemGroup Condition="'$(TargetFramework)' == 'netstandard1.4'">
+        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="1.1.2"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="1.1.2"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="1.1.1"/>
+
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\ATag.Core\ATag.Core.csproj"/>
+    </ItemGroup>
+
+</Project>
