+++ src/ATag.EntityFrameworkCore/DbEntityConfiguration.cs
