+++ tests/ATag.Tests/ATag.Tests.csproj
@@ -1,0 +1,43 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup Label="Configuration" Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+        <OutputType>Library</OutputType>
+    </PropertyGroup>
+
+    <PropertyGroup>
+        <TargetFramework>netcoreapp2.0</TargetFramework>
+        <ApplicationIcon/>
+        <OutputTypeEx>library</OutputTypeEx>
+        <StartupObject/>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <None Remove="appsettings.json"/>
+    </ItemGroup>
+
+    <ItemGroup>
+        <Content Include="appsettings.json">
+            <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+        </Content>
+    </ItemGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="2.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.0.0"/>
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0"/>
+        <PackageReference Include="xunit" Version="2.3.0-beta5-build3769"/>
+        <PackageReference Include="xunit.runner.visualstudio" Version="2.3.0-beta5-build3769"/>
+        <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.0-beta2-build3683"/>
+        <DotNetCliToolReference Include="Microsoft.EntityFrameworkCore.Tools.DotNet" Version="1.0.0"/>
+    </ItemGroup>
+
+    <ItemGroup>
+        <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}"/>
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\..\src\ATag.EntityFrameworkCore\ATag.EntityFrameworkCore.csproj"/>
+    </ItemGroup>
+
+</Project>
